Add ICPSHC24-10EU-IL-1 (10W LED driver/dimmer)
After getting a 10 Watt transformer I opened it for documentation purposes.
Find the teardown and product information.

The power control unit appears to be identical to the ICPSHC24-30EU-IL-1
as can be seen in PR #7. Further investigation is needed and likely
deblobbing/soldering to confirm some traces.

U1 does not seem to be controlled by the rf-module, but is very likely to be a buck converter (24 V -> 3.3 V).

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>