schematic: ICPSLC24-10NA PWM module
With the teardown of the ICPSLC24-10NA module complete, a KiCAD
schematic has been drawn to understand its exact function.

A PCB design has been excluded at this moment.

Note: The buck regulator is not part of the standard library, but
is on its way upstream to the digi-key kicad repo and likely the
default kicad repo as well.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>