(Re-)create parts based on datasheets
This commit creates the ICC-1 poly modules based on their limited
datasheet [0] information. The datasheet properly lists pin numbers
and acurate dimensions, such as the most important dimension,
pin locations.

The parts are added as two variants, one a pure SMD mount and one where the
SMD pats have received throug-hole mounts. This allows for via-less routing
and soldering pins to the module and insert it into the board that way.

The wings3d module was created using the existing stl file. More details could
be added, such as pads and holes to make it more accurate.

[0] https://fccid.io/FHO-ICC-1/User-Manual/User-Manual-3055710.pdf
Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>